Split wait helper extensions into two files: one for unit tests, another for UIKit based integration  tests
